pyproject.toml
[build-system]
requires = ["scikit-build-core", "pybind11"]
build-backend = "scikit_build_core.build"

[project]
name = "mad-icp"
version = "0.0.8"
description = "It Is All About Matching Data -- Robust and Informed LiDAR Odometry"
readme = "README.md"
authors = [
    { name = "Simone Ferrari", email = "s.ferrari@diag.uniroma1.it" },
    { name = "Luca Di Giammarino", email = "digiammarino@diag.uniroma1.it" },
    { name = "Leonardo Brizi", email = "brizi@diag.uniroma1.it" },
    { name = "Emanuele Giacomini", email = "giacomini@diag.uniroma1.it" },
]
keywords = ["LiDAR"]
requires-python = ">=3.8"

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: Unix",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
]
dependencies = [
    "PyYAML==6.0.1",
    "numpy==1.23.5",
    "rosbags==0.9.23",
    "open3d",
    "matplotlib==3.1.2",
    "typer==0.12.3",
    "natsort==8.4.0",
    "rich==13.7.1",
]

[project.scripts]
mad_icp = "mad_icp.apps.mad_icp:run"

[project.urls]
Homepage = "https://github.com/rvp-group/mad-icp"

[tool.scikit-build]
build-dir = "build/{wheel_tag}"
cmake.source-dir = "mad_icp/"
cmake.verbose = false
cmake.version = ">=3.8"
wheel.install-dir = "mad_icp/src/pybind"
